Implemented changes for new DriveTrain Modules

+++ DriveTrainMK4/PID.cpp
@@ -2,7 +2,7 @@
 
 
 // Constructors
-PID::PID(double _Sp, double _Dt, double _Kp, double _Ki, double _Kd):
+PID::PID(uint8_t _Sp, double _Dt, double _Kp, double _Ki, double _Kd):
   Sp(_Sp), Dt(_Dt), Integral(0), pErr(0), Kp(_Kp), Ki(_Ki), Kd(_Kd)
 {
   // intentionally blank
